+++ operator/internal/manifests/internal/config/build_test.go
@@ -28,6 +28,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -50,12 +56,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -200,9 +204,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -264,6 +269,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -286,12 +297,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -453,9 +462,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -527,9 +537,10 @@ func TestBuild_ConfigAndRuntimeConfig_CreateLokiConfigFailed(t *testing.T) {
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -590,6 +601,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -612,12 +629,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -816,9 +831,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -927,6 +943,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -949,12 +971,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -1153,9 +1173,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -1265,6 +1286,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -1287,12 +1314,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -1504,9 +1529,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -1633,6 +1659,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -1658,12 +1690,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -1851,9 +1881,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -1922,6 +1953,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -1944,12 +1981,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -2174,9 +2209,10 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",
@@ -2320,6 +2356,12 @@ common:
       secret_access_key: test123
       s3forcepathstyle: true
   compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_port: 9095
 compactor:
   compaction_interval: 2h
   working_directory: /tmp/loki/compactor
@@ -2356,12 +2398,10 @@ ingester:
   flush_op_timeout: 10m
   lifecycler:
     final_sleep: 0s
-    heartbeat_period: 5s
     join_after: 30s
     num_tokens: 512
     ring:
       replication_factor: 1
-      heartbeat_timeout: 1m
   max_chunk_age: 2h
   max_transfer_retries: 0
   wal:
@@ -2581,9 +2621,253 @@ overrides:
 			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
 			Port: 9095,
 		},
-		GossipRing: Address{
-			FQDN: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
-			Port: 7946,
+		GossipRing: GossipRing{
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
+		},
+		Querier: Address{
+			Protocol: "http",
+			FQDN:     "loki-querier-http-lokistack-dev.default.svc.cluster.local",
+			Port:     3100,
+		},
+		IndexGateway: Address{
+			FQDN: "loki-index-gateway-grpc-lokistack-dev.default.svc.cluster.local",
+			Port: 9095,
+		},
+		StorageDirectory: "/tmp/loki",
+		MaxConcurrent: MaxConcurrent{
+			AvailableQuerierCPUCores: 2,
+		},
+		WriteAheadLog: WriteAheadLog{
+			Directory:             "/tmp/wal",
+			IngesterMemoryRequest: 5000,
+		},
+		ObjectStorage: storage.Options{
+			SharedStore: lokiv1.ObjectStorageSecretS3,
+			S3: &storage.S3StorageConfig{
+				Endpoint:        "http://test.default.svc.cluster.local.:9000",
+				Region:          "us-east",
+				Buckets:         "loki",
+				AccessKeyID:     "test",
+				AccessKeySecret: "test123",
+			},
+			Schemas: []lokiv1.ObjectStorageSchema{
+				{
+					Version:       lokiv1.ObjectStorageSchemaV11,
+					EffectiveDate: "2020-10-01",
+				},
+			},
+		},
+		EnableRemoteReporting: true,
+	}
+	cfg, rCfg, err := Build(opts)
+	require.NoError(t, err)
+	require.YAMLEq(t, expCfg, string(cfg))
+	require.YAMLEq(t, expRCfg, string(rCfg))
+}
+
+func TestBuild_ConfigAndRuntimeConfig_WithHashRingSpec(t *testing.T) {
+	expCfg := `
+---
+auth_enabled: true
+chunk_store_config:
+  chunk_cache_config:
+    embedded_cache:
+      enabled: true
+      max_size_mb: 500
+common:
+  storage:
+    s3:
+      s3: http://test.default.svc.cluster.local.:9000
+      bucketnames: loki
+      region: us-east
+      access_key_id: test
+      secret_access_key: test123
+      s3forcepathstyle: true
+  compactor_grpc_address: loki-compactor-grpc-lokistack-dev.default.svc.cluster.local:9095
+  ring:
+    kvstore:
+      store: memberlist
+    heartbeat_period: 5s
+    heartbeat_timeout: 1m
+    instance_addr: ${HASH_RING_INSTANCE_ADDR}
+    instance_port: 9095
+compactor:
+  compaction_interval: 2h
+  working_directory: /tmp/loki/compactor
+frontend:
+  tail_proxy_url: http://loki-querier-http-lokistack-dev.default.svc.cluster.local:3100
+  compress_responses: true
+  max_outstanding_per_tenant: 256
+  log_queries_longer_than: 5s
+frontend_worker:
+  frontend_address: loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local:9095
+  grpc_client_config:
+    max_send_msg_size: 104857600
+  match_max_concurrent: true
+ingester:
+  chunk_block_size: 262144
+  chunk_encoding: snappy
+  chunk_idle_period: 1h
+  chunk_retain_period: 5m
+  chunk_target_size: 2097152
+  flush_op_timeout: 10m
+  lifecycler:
+    final_sleep: 0s
+    join_after: 30s
+    num_tokens: 512
+    ring:
+      replication_factor: 1
+  max_chunk_age: 2h
+  max_transfer_retries: 0
+  wal:
+    enabled: true
+    dir: /tmp/wal
+    replay_memory_ceiling: 2500
+ingester_client:
+  grpc_client_config:
+    max_recv_msg_size: 67108864
+  remote_timeout: 1s
+# NOTE: Keep the order of keys as in Loki docs
+# to enable easy diffs when vendoring newer
+# Loki releases.
+# (See https://grafana.com/docs/loki/latest/configuration/#limits_config)
+#
+# Values for not exposed fields are taken from the grafana/loki production
+# configuration manifests.
+# (See https://github.com/grafana/loki/blob/main/production/ksonnet/loki/config.libsonnet)
+limits_config:
+  ingestion_rate_strategy: global
+  ingestion_rate_mb: 4
+  ingestion_burst_size_mb: 6
+  max_label_name_length: 1024
+  max_label_value_length: 2048
+  max_label_names_per_series: 30
+  reject_old_samples: true
+  reject_old_samples_max_age: 168h
+  creation_grace_period: 10m
+  enforce_metric_name: false
+  # Keep max_streams_per_user always to 0 to default
+  # using max_global_streams_per_user always.
+  # (See https://github.com/grafana/loki/blob/main/pkg/ingester/limiter.go#L73)
+  max_streams_per_user: 0
+  max_line_size: 256000
+  max_entries_limit_per_query: 5000
+  max_global_streams_per_user: 0
+  max_chunks_per_query: 2000000
+  max_query_length: 721h
+  max_query_parallelism: 32
+  max_query_series: 500
+  cardinality_limit: 100000
+  max_streams_matchers_per_query: 1000
+  max_cache_freshness_per_query: 10m
+  per_stream_rate_limit: 3MB
+  per_stream_rate_limit_burst: 15MB
+  split_queries_by_interval: 30m
+  query_timeout: 1m
+memberlist:
+  abort_if_cluster_join_fails: true
+  bind_port: 7946
+  join_members:
+    - loki-gossip-ring-lokistack-dev.default.svc.cluster.local:7946
+  max_join_backoff: 1m
+  max_join_retries: 10
+  min_join_backoff: 1s
+querier:
+  engine:
+    max_look_back_period: 30s
+    timeout: 3m
+  extra_query_delay: 0s
+  max_concurrent: 2
+  query_ingesters_within: 3h
+  tail_max_duration: 1h
+query_range:
+  align_queries_with_step: true
+  cache_results: true
+  max_retries: 5
+  results_cache:
+    cache:
+      embedded_cache:
+        enabled: true
+        max_size_mb: 500
+  parallelise_shardable_queries: true
+schema_config:
+  configs:
+    - from: "2020-10-01"
+      index:
+        period: 24h
+        prefix: index_
+      object_store: s3
+      schema: v11
+      store: boltdb-shipper
+server:
+  graceful_shutdown_timeout: 5s
+  grpc_server_min_time_between_pings: '10s'
+  grpc_server_ping_without_stream_allowed: true
+  grpc_server_max_concurrent_streams: 1000
+  grpc_server_max_recv_msg_size: 104857600
+  grpc_server_max_send_msg_size: 104857600
+  http_listen_port: 3100
+  http_server_idle_timeout: 120s
+  http_server_write_timeout: 1m
+  log_level: info
+storage_config:
+  boltdb_shipper:
+    active_index_directory: /tmp/loki/index
+    cache_location: /tmp/loki/index_cache
+    cache_ttl: 24h
+    resync_interval: 5m
+    shared_store: s3
+    index_gateway_client:
+      server_address: dns:///loki-index-gateway-grpc-lokistack-dev.default.svc.cluster.local:9095
+tracing:
+  enabled: false
+analytics:
+  reporting_enabled: true
+`
+	expRCfg := `
+---
+overrides:
+`
+	opts := Options{
+		Stack: lokiv1.LokiStackSpec{
+			ReplicationFactor: 1,
+			Limits: &lokiv1.LimitsSpec{
+				Global: &lokiv1.LimitsTemplateSpec{
+					IngestionLimits: &lokiv1.IngestionLimitSpec{
+						IngestionRate:             4,
+						IngestionBurstSize:        6,
+						MaxLabelNameLength:        1024,
+						MaxLabelValueLength:       2048,
+						MaxLabelNamesPerSeries:    30,
+						MaxGlobalStreamsPerTenant: 0,
+						MaxLineSize:               256000,
+					},
+					QueryLimits: &lokiv1.QueryLimitSpec{
+						MaxEntriesLimitPerQuery: 5000,
+						MaxChunksPerQuery:       2000000,
+						MaxQuerySeries:          500,
+						QueryTimeout:            "1m",
+					},
+				},
+			},
+		},
+		Namespace: "test-ns",
+		Name:      "test",
+		Compactor: Address{
+			FQDN: "loki-compactor-grpc-lokistack-dev.default.svc.cluster.local",
+			Port: 9095,
+		},
+		FrontendWorker: Address{
+			FQDN: "loki-query-frontend-grpc-lokistack-dev.default.svc.cluster.local",
+			Port: 9095,
+		},
+		GossipRing: GossipRing{
+			InstanceAddr:         "${HASH_RING_INSTANCE_ADDR}",
+			InstancePort:         9095,
+			BindPort:             7946,
+			MembersDiscoveryAddr: "loki-gossip-ring-lokistack-dev.default.svc.cluster.local",
 		},
 		Querier: Address{
 			Protocol: "http",